Extract title frame source
